code updated accoding to review
