+++ src/fetchers/makesFetcher.ts
@@ -1,4 +1,4 @@
-import { axiosInstance } from ".";
+import { axiosInstance } from "./axiosInstance";
 
 export async function makesFetcher(url: string) {
     const data = await axiosInstance.get(url);
